Adjust meson dependency to freetype2, instead of freetype
Not fully checked, but my new laptop build has the freetype2 package and
failed to build with plain freetype, despite me knowing the latter was
needed for some other distributions.
Potential for this to come up in build failures on some systems.

May simply try both with logic handling if it seems worthwhile.